tests/wpt/tests/css/css-animations/parsing/animation-trigger-range-shorthand.html

<!DOCTYPE html>
<title>animation-trigger-range shorthand</title>
<!-- TODO(crbug.com/390314945): Replace with spec link when the spec lands. -->
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8942">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="target" style="font-size:10px"></div>
<script>
test_valid_value("animation-trigger-range", "normal");
test_valid_value("animation-trigger-range", "normal normal", "normal");
test_valid_value("animation-trigger-range", "cover");
test_valid_value("animation-trigger-range", "contain");
test_valid_value("animation-trigger-range", "entry");
test_valid_value("animation-trigger-range", "entry-crossing");
test_valid_value("animation-trigger-range", "exit");
test_valid_value("animation-trigger-range", "exit-crossing");
test_valid_value("animation-trigger-range", "entry, exit");

test_valid_value("animation-trigger-range", "entry 0% entry 100%", "entry");
test_valid_value("animation-trigger-range", "entry-crossing 0% entry-crossing 100%",
                 "entry-crossing");
test_valid_value("animation-trigger-range", "exit 0% exit 100%", "exit");
test_valid_value("animation-trigger-range", "exit-crossing 0% exit-crossing 100%",
                 "exit-crossing");
test_valid_value("animation-trigger-range", "cover 0% cover 100%", "cover");
test_valid_value("animation-trigger-range", "contain 0% contain 100%", "contain");
test_valid_value("animation-trigger-range",
                 "entry calc(10% - 10%) entry calc(50% + 50%)", "entry");
test_valid_value("animation-trigger-range", "cover 50%");
test_valid_value("animation-trigger-range", "contain 50%");
test_valid_value("animation-trigger-range", "entry 50%");
test_valid_value("animation-trigger-range", "entry-crossing 50%");
test_valid_value("animation-trigger-range", "exit 50%");
test_valid_value("animation-trigger-range", "exit-crossing 50%");
test_valid_value("animation-trigger-range", "entry 50px exit 100px");
test_valid_value("animation-trigger-range", "exit calc(10% + 50px)");

test_valid_value("animation-trigger-range", "entry 50% exit 50%");
test_valid_value("animation-trigger-range",
                 "cover 50% entry 50%, contain 50% exit 50%");
test_valid_value("animation-trigger-range", "50% exit 50%");
test_valid_value("animation-trigger-range", "normal 100px");
test_valid_value("animation-trigger-range", "100px");
test_valid_value("animation-trigger-range", "100px normal", "100px");
test_valid_value("animation-trigger-range", "10% normal", "10%");

test_computed_value("animation-trigger-range", "normal");
test_computed_value("animation-trigger-range", "normal normal", "normal");
test_computed_value("animation-trigger-range", "cover");
test_computed_value("animation-trigger-range", "contain");
test_computed_value("animation-trigger-range", "entry");
test_computed_value("animation-trigger-range", "entry-crossing");
test_computed_value("animation-trigger-range", "exit");
test_computed_value("animation-trigger-range", "exit-crossing");
test_computed_value("animation-trigger-range", "entry, exit");

test_computed_value("animation-trigger-range", "entry 0% entry 100%", "entry");
test_computed_value("animation-trigger-range", "entry-crossing 0% entry-crossing 100%",
                 "entry-crossing");
test_computed_value("animation-trigger-range", "exit 0% exit 100%", "exit");
test_computed_value("animation-trigger-range", "exit-crossing 0% exit-crossing 100%",
                 "exit-crossing");
test_computed_value("animation-trigger-range", "cover 0% cover 100%", "cover");
test_computed_value("animation-trigger-range", "contain 0% contain 100%", "contain");
test_computed_value("animation-trigger-range",
                 "entry calc(10% - 10%) entry calc(50% + 50%)", "entry");
test_computed_value("animation-trigger-range", "cover 50%");
test_computed_value("animation-trigger-range", "contain 50%");
test_computed_value("animation-trigger-range", "entry 50%");
test_computed_value("animation-trigger-range", "entry-crossing 50%");
test_computed_value("animation-trigger-range", "exit 50%");
test_computed_value("animation-trigger-range", "exit-crossing 50%");
test_computed_value("animation-trigger-range", "entry 50px exit 100px");
test_computed_value("animation-trigger-range", "exit calc(10% + 50px)");

test_computed_value("animation-trigger-range", "entry 50% exit 50%");
test_computed_value("animation-trigger-range",
                 "cover 50% entry 50%, contain 50% exit 50%");

test_computed_value("animation-trigger-range", "entry 10em exit 20em", "entry 100px exit 200px");
test_computed_value("animation-trigger-range", "10em exit 20em", "100px exit 200px");
test_computed_value("animation-trigger-range", "normal 100px");
test_computed_value("animation-trigger-range", "100px");
test_computed_value("animation-trigger-range", "100px normal", "100px");
test_computed_value("animation-trigger-range", "10% normal", "10%");
test_computed_value("animation-trigger-range", "10% calc(70% + 10% * sign(100em - 1px))", "10% 80%");

test_invalid_value("animation-trigger-range", "entry 50% 0s", "entry 50%");
test_invalid_value("animation-trigger-range", "0s entry 50%");
test_invalid_value("animation-trigger-range", "1s");
test_invalid_value("animation-trigger-range", "-1s");
test_invalid_value("animation-trigger-range", "1s 2s");
test_invalid_value("animation-trigger-range", "1s, 2s");
test_invalid_value("animation-trigger-range", "1s 2s, 3s");
test_invalid_value("animation-trigger-range", "1s, 2s 3s");
test_invalid_value("animation-trigger-range", "1s, 2s, 3s");
test_invalid_value("animation-trigger-range", "1s 2s 3s");
test_invalid_value("animation-trigger-range", "0s, 1s 2s 3s");
test_invalid_value("animation-trigger-range", "1s / 2s");
test_invalid_value("animation-trigger-range", "1s, 2px");
test_invalid_value("animation-trigger-range", "#ff0000");
test_invalid_value("animation-trigger-range", "red");
test_invalid_value("animation-trigger-range", "thing");
test_invalid_value("animation-trigger-range", "thing 0%");
test_invalid_value("animation-trigger-range", "thing 42%");
test_invalid_value("animation-trigger-range", "thing 100%");
test_invalid_value("animation-trigger-range", "thing 100px");
test_invalid_value("animation-trigger-range", "100% thing");

test_shorthand_value('animation-trigger-range', 'normal', {
  'animation-trigger-range-start': 'normal',
  'animation-trigger-range-end': 'normal',
});

test_shorthand_value('animation-trigger-range', 'normal normal', {
  'animation-trigger-range-start': 'normal',
  'animation-trigger-range-end': 'normal',
});

test_shorthand_value('animation-trigger-range', 'normal entry 100%', {
  'animation-trigger-range-start': 'normal',
  'animation-trigger-range-end': 'entry',
});

test_shorthand_value('animation-trigger-range', 'normal entry 10%', {
  'animation-trigger-range-start': 'normal',
  'animation-trigger-range-end': 'entry 10%',
});

test_shorthand_value('animation-trigger-range', 'cover', {
  'animation-trigger-range-start': 'cover',
  'animation-trigger-range-end': 'cover',
});

test_shorthand_value('animation-trigger-range', 'contain', {
  'animation-trigger-range-start': 'contain',
  'animation-trigger-range-end': 'contain',
});

test_shorthand_value('animation-trigger-range', 'contain 100% contain 0%', {
  'animation-trigger-range-start': 'contain 100%',
  'animation-trigger-range-end': 'contain 0%',
});

test_shorthand_value('animation-trigger-range', 'entry 10% exit 20%', {
  'animation-trigger-range-start': 'entry 10%',
  'animation-trigger-range-end': 'exit 20%',
});

test_shorthand_value('animation-trigger-range', 'entry calc(10% + 10px) exit 20px', {
  'animation-trigger-range-start': 'entry calc(10% + 10px)',
  'animation-trigger-range-end': 'exit 20px',
});

test_shorthand_value('animation-trigger-range', 'entry, exit', {
  'animation-trigger-range-start': 'entry, exit',
  'animation-trigger-range-end': 'entry, exit',
});

test_shorthand_value('animation-trigger-range', 'entry 0%, exit', {
  'animation-trigger-range-start': 'entry, exit',
  'animation-trigger-range-end': 'entry, exit',
});

test_shorthand_value('animation-trigger-range', 'exit calc(10% + 50px)', {
  'animation-trigger-range-start': 'exit calc(10% + 50px)',
  'animation-trigger-range-end': 'exit',
});
test_shorthand_value('animation-trigger-range', '100px', {
  'animation-trigger-range-start': '100px',
  'animation-trigger-range-end': 'normal',
});
test_shorthand_value('animation-trigger-range', '10%', {
  'animation-trigger-range-start': '10%',
  'animation-trigger-range-end': 'normal',
});
</script>
